+++ src/main/java/roomescape/common/auth/LoginCustomerResolver.java
@@ -1,4 +1,4 @@
-package roomescape.config;
+package roomescape.common.auth;
 
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
