+++ src/test/java/roomescape/MissionStepTest.java
@@ -19,7 +19,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.annotation.DirtiesContext;
 import roomescape.common.Role;
-import roomescape.common.util.JwtProvider;
+import roomescape.common.auth.JwtProvider;
 import roomescape.dao.ReservationTimeDao;
 import roomescape.dao.ThemeDao;
 import roomescape.controller.reservation.dto.ReservationResponseDto;
