+++ src/test/java/roomescape/controller/ReservationControllerTest.java
@@ -10,7 +10,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.DirtiesContext;
 import roomescape.common.Role;
-import roomescape.common.util.JwtProvider;
+import roomescape.common.auth.JwtProvider;
 import roomescape.model.Member;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
