+++ src/test/java/roomescape/controller/ReservationTimeControllerTest.java
@@ -12,7 +12,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.annotation.DirtiesContext;
 import roomescape.common.Role;
-import roomescape.common.util.JwtProvider;
+import roomescape.common.auth.JwtProvider;
 import roomescape.model.Member;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
